Scale visualisation based on the current number of lanes. This way size of cars and position of lanes stays the same as the number of lanes is changed in the web visualisation.
Simple data collecter added for checking the velocity of car 1.
Collected data is shown in a graph in the web visualisation.

Slider for changing the maximum speed (in km/h) added.

Easily switch between timing schedule:
random or staged (update velocities, move cars).

Lane switching is broken (and disabled) in this commit.